finishing the service classes

+++ src/main/java/com/spotily/app/song/SongService.java
@@ -1,7 +1,9 @@
 package com.spotily.app.song;
 
+import com.spotily.app.exception.ResourceNotFound;
 import org.springframework.stereotype.Service;
 
+import java.lang.module.ResolutionException;
 import java.util.Optional;
 
 @Service
@@ -16,12 +18,19 @@ public class SongService {
     public void deleteSong(int id) {
         Optional<Song> songOptional = songDAO.deleteSong(id);
 
-
-
+        if(songOptional.isEmpty()){
+            throw new ResourceNotFound("Song not found, field can't be empty");
+        }
         songDAO.deleteSong(id);
     }
 
     public void editSongDetails(int id, Song song) {
+
+        Optional<Song> songOptional = songDAO.editSongDetails(id, song);
+
+        if (songOptional.isEmpty()){
+            throw new ResourceNotFound("Song not found, field can't be empty");
+        }
         songDAO.editSongDetails(id, song);
     }
 }
